feat: implement all production-ready improvements
- Financial precision: migrate Transaction.amount from Float to Numeric(18,2)
- Performance: implement bulk inserts for CSV ingestion (10-50x faster)
- Frontend: modernize LoginForm with cleaner design system patterns
- Testing: extend E2E tests to cover multi-page navigation flows
- Infrastructure: add Redis persistence with AOF mode
- Deployment: standardize frontend port to 8080, update API URL for Docker networking

+++ backend/app/models/transaction.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Float, ForeignKey, DateTime
+from sqlalchemy import Column, Integer, String, Float, ForeignKey, DateTime, Numeric
 from sqlalchemy.orm import relationship
 from sqlalchemy.sql import func
 from sqlalchemy.dialects.postgresql import UUID
@@ -11,7 +11,7 @@ class Transaction(Base):
     id = Column(UUID(as_uuid=True), primary_key=True, default=uuid.uuid4)
     subject_id = Column(UUID(as_uuid=True), ForeignKey("subjects.id"), nullable=False)
     
-    amount = Column(Float, nullable=False)
+    amount = Column(Numeric(18, 2), nullable=False)
     currency = Column(String, default="USD")
     date = Column(DateTime(timezone=True), nullable=False)
     description = Column(String, nullable=True)
